reminder api added

+++ lib/domain/model/norms/norms.dart
@@ -5,7 +5,7 @@ part 'norms.g.dart';
 
 @freezed
 abstract class NormsRequest with _$NormsRequest {
-  const factory NormsRequest({required String metric, required int value}) = _NormsRequest;
+  const factory NormsRequest({required String metric, required double value}) = _NormsRequest;
 
   factory NormsRequest.fromJson(Map<String, dynamic> json) => _$NormsRequestFromJson(json);
 }
